Inline pagination + response wrappers into client module
